requirements.txt
Flask==2.3.2
python_jose==3.3.0
Requests==2.31.0
Flask-Cors==3.0.10
